requirements.txt
python-dotenv
pandas
streamlit
psycopg2-binary